Commit de mudanças na organização da estrutura referente ao dashboard e profile. Corrigido bug dos nomes das tabelas imagem e profile
