+++ apps/autenticacao/apps.py
@@ -1,6 +1,5 @@
 from django.apps import AppConfig
 
-
 class AutenticacaoConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'autenticacao'
