+++ apps/autenticacao/signals.py
@@ -1,8 +1,7 @@
 from django.db.models.signals import post_save #Import a post_save signal when a user is created
 from django.contrib.auth.models import User # Import the built-in User model, which is a sender
 from django.dispatch import receiver # Import the receiver
-from .models import Profile
-
+from dashboard.models import Profile
 
 @receiver(post_save, sender=User) # The receiver decorator takes the signal and the sender as arguments
 def create_profile(sender, instance, created, **kwargs): # The receiver function takes the arguments of the signal
