+++ apps/dashboard/admin.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
-from .models import Imagem
+from .models import Imagem, Profile
 
+admin.site.register(Profile)
 admin.site.register(Imagem)
