+++ apps/dashboard/forms.py
@@ -1,0 +1,7 @@
+from django import forms
+from .models import Profile
+
+class ProfileForm(forms.ModelForm):
+    class Meta:
+        model = Profile
+        fields = ['foto_de_perfil', 'descricao_de_perfil']
