+++ apps/dashboard/models.py
@@ -9,3 +9,19 @@ class Imagem(models.Model):
 
     def __UUID__(self): 
         return self.id
+    
+    class Meta:
+        verbose_name = "Imagem"
+        verbose_name_plural = "Imagens"
+
+class Profile(models.Model):
+    usuario = models.OneToOneField(User, on_delete=models.CASCADE)
+    foto_de_perfil = models.ImageField(upload_to='static/media/fotos_de_perfil', default='static/media/fotos_de_perfil/default.jpg')
+    descricao_de_perfil = models.TextField(max_length=1000, blank=True)
+
+    def __str__(self):
+        return self.usuario.get_full_name()
+    
+    class Meta:
+        verbose_name = "Perfil"
+        verbose_name_plural = "Perfis"
